Restructured test packages.
