+++ test/ch/epfl/leb/sass/models/psfs/internal/GibsonLanniPSFTest.java
@@ -15,7 +15,7 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package ch.epfl.leb.sass.simulator.generators.realtime.psfs;
+package ch.epfl.leb.sass.models.psfs.internal;
 
 import ch.epfl.leb.sass.models.psfs.PSF;
 import ch.epfl.leb.sass.models.psfs.internal.GibsonLanniPSF;
